Added a decision for the USSR to blow the Dnieper Dam
The decision requires the USSR to have lost Kiev and hold Zaporozhe, or have lost one of Zaporozhe and Dnipropetrovsk and control the other.

The decision damages the dam (-90% malus to repair rate) in state 200 (Zaporozhe) and floods 196 (Kherson) and 200 with the same modifier as the Yellow River Flood:

Manpower -50%
Speed -25%
Strat Redeploy Disabled
Local Supply Impact +50%
Local Available Resources -25%

As well as damaging the infrastructure and mil + civ factories and AA and airbases in 196 and 200

Also added a fun little news event so everyone can see that the USSR has demonstrated their resolve by destroying the very symbol of the five-year plan.