Fixed all the little italy problems, made the decision to annex ETH delete the cores
Removes the "2nd Italo-Ethiopian War" category and "Withdraw from Ethiopia" decision.

Makes the decision for ITA to annex ETH remove the cores for ETH, so you can't release it even if you wanted to do so.
The ETH_decision_categories text file is included to make sure the game doesn't look into the vanilla file.