Fixes the special forces path for Canada
Changes the 2x100% marines bonus to 2x100% special forces bonus, because fuck you, nobody is using the 100% on the second marine tech, what are you smoking.
Fixes the doctrine cost reduction for special forces doctrine, kept the value exactly as it was (no balance change).

Nobody will ever pick this path anyway so it doesn't matter.